feat: add pop-up login - partial
