+++ src/views/Home/index.tsx
@@ -1,33 +1,40 @@
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
 
 import {
-  Tile,
   Button,
   FlexGrid,
   Column,
   Row,
   Stack
 } from '@carbon/react';
-import { ArrowRight } from '@carbon/icons-react';
 import { useKeycloak } from '@react-keycloak/web';
-import { kcClientRoles, kcUserDisplayName, kcIdentityProvider } from '../../service/AuthService';
+import { KeycloakLoginOptions } from 'keycloak-js';
 
 const Home = () => {
-  const navigate = useNavigate();
-  const { keycloak, initialized } = useKeycloak();
+  const { keycloak } = useKeycloak();
 
-  const goToTable = () => {
-    navigate('/table');
+  const createLoginUrl = (provider: 'idir' | 'bceid'): string => {
+    const idpHint = provider === 'idir' ? 'idir' : 'bceid-business';
+
+    const loginOptions: KeycloakLoginOptions = {
+      redirectUri: `${window.location.origin}/*`,
+      idpHint
+    };
+
+    return keycloak.createLoginUrl(loginOptions);
   };
 
-  const goToForm = () => {
-    navigate('/form');
+  const loginWithIdir = () => {
+    const idirUrl = createLoginUrl('idir');
+    console.log('IDIR! ', idirUrl);
+    const windowFeatures = 'width=450,height=600,popup=true';
+    window.open(idirUrl, 'loginWindow', windowFeatures);
   };
 
-  const provider = (): string => ` With ${kcIdentityProvider(keycloak, initialized)}`;
-  const welcomeMsg = (): string => `Welcome ${kcUserDisplayName(keycloak, initialized)}`;
-  const showRoles = (): string => `Your roles: ${kcClientRoles(keycloak, initialized)}`;
+  const loginWithBCeID = () => {
+    const bceidUrl = createLoginUrl('bceid');
+    console.log('BCeID! ', bceidUrl);
+  };
 
   return (
     <FlexGrid container="true" spacing={4}>
@@ -35,7 +42,7 @@ const Home = () => {
         <Row>
           <Column sm={4}>
             <Stack gap={3}>
-              <h1 data-testid="home-title">NR Front End Testing App</h1>
+              <h1 data-testid="home-title">NR Sample App</h1>
               <p data-testid="home-desc">
                 This is a sample application with the purpose of checking BC Gov&apos;s design
                 system, themes and react components. The app will also be used to validate
@@ -46,80 +53,21 @@ const Home = () => {
         </Row>
         <Row>
           <Column sm={4} md={4}>
-            <Tile data-testid="card-form">
-              <h3 data-testid="card-form__title">Sample User Form</h3>
-              <br />
-              <p data-testid="card-form__desc">
-                This is a simple form to validate the user&apos;s inputs.
-              </p>
-              <br />
-              {keycloak?.authenticated && (
-                <Button
-                  onClick={goToForm}
-                  size="md"
-                  renderIcon={ArrowRight}
-                  data-testid="card-form__button"
-                >
-                  Go!
-                </Button>
-              )}
-              {!keycloak?.authenticated && (
-                <>
-                  <br />
-                  <p>You&apos;re now allowed to click the button, sorry.</p>
-                </>
-              )}
-            </Tile>
-          </Column>
-          <Column sm={4} md={4}>
-            <Tile data-testid="card-table">
-              <h3 data-testid="card-table__title">Sample Search</h3>
-              <br />
-              <p data-testid="card-table__desc">
-                This is a simple table showing results of a simulated search.
-              </p>
-              <br />
-              <Button
-                onClick={goToTable}
-                size="md"
-                renderIcon={ArrowRight}
-                data-testid="card-table__button"
-              >
-                Go!
-              </Button>
-            </Tile>
-          </Column>
-        </Row>
-        <Row>
-          <Column>
-            <Tile data-testid="card-authentication">
-              <h3 data-testid="card-authentication__title">Authentication</h3>
-              <br />
-              {!!keycloak?.authenticated && (
-                <div>
-                  <p data-testid="card-table__desc">
-                    You are authenticated!
-                    { provider() }
-                  </p>
-                  <p data-testid="card-table__desc">
-                    { welcomeMsg() }
-                  </p>
-                  <p>
-                    { showRoles() }
-                  </p>
-                </div>
-              )}
-              {!keycloak?.authenticated && (
-                <div>
-                  <p data-testid="card-table__desc">
-                    You are not authenticated!
-                  </p>
-                  <p data-testid="card-table__desc">
-                    Please, hit the Login button at the upper right corner!
-                  </p>
-                </div>
-              )}
-            </Tile>
+            <Button
+              onClick={loginWithIdir}
+              size="lg"
+              data-testid="card-table__button"
+            >
+              Login with IDIR
+            </Button>
+            &nbsp;
+            <Button
+              onClick={loginWithBCeID}
+              size="lg"
+              data-testid="card-table__button"
+            >
+              Login with Business BCeID
+            </Button>
           </Column>
         </Row>
       </Stack>
